Update French UI to 1/22/13 POT file
Adds translations for all UI strings up to what was in the sources
on 1/22/13.

Fixes a problem where previous .po file was edited with
PoEdit 1.5.4, and that version messes up the line breaks for the msgid
strings.  THIS VERSION OF POEDIT SHOULD NOT BE USED!  The version I'm
checking in now was edited with PoEdit 1.4.6, and that keeps the line
endings stable.

100% translated.

Change-Id: I4ccf6d3fafff7b8c2c00a11451f0958dd5bf67c5