LT-2352: Revert part of the change for import message
The best way to get the new version of the import message
translated is to leave the relevant strings as they had
been.  When a new POT file is generated, these will come
up as "need work".  The translator will be able to retain
the old translation and add to it.

Change-Id: I57865a26f896aca1fa10c5796b6ad79d005512b7