Unwrap lines
This commit makes no content changes.  It reverses the line
wrapping that was introduced in the Mar 30 2017 commit.  Now
a diff can be done with the commit prior to that, and the
results are more useful for seeing what changed.  (They still
have comments in different orders; I don't know how to prevent
that.)

Change-Id: I6ec61806f743603a8c508a34e8b65a5149b4d70d